+++ src/mechanisms/plain/mechinfo.rs
@@ -24,6 +24,8 @@ pub static PLAIN: Mechanism = Mechanism {
 pub(super) enum PlainError {
     #[error("The given value contains a NULL-byte")]
     ContainsNull,
+    #[error("authid and password must not be empty")]
+    Empty,
     #[error("invalid format, expected three strings separated by two NULL-bytes")]
     BadFormat,
     #[error("authzid is invalid UTF-8: {0}")]