+++ src/mechanisms/anonymous/client.rs
@@ -1,14 +1,9 @@
 use crate::context::EmptyProvider;
 use crate::error::SessionError;
 use crate::mechanism::Authentication;
-use crate::property::Property;
 use crate::session::{MechanismData, State};
 use std::io::Write;
-
-pub struct AnonymousToken;
-impl Property<'_> for AnonymousToken {
-    type Value = str;
-}
+use super::AnonymousToken;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Anonymous;
@@ -19,13 +14,66 @@ impl Authentication for Anonymous {
         _input: Option<&[u8]>,
         writer: &mut dyn Write,
     ) -> Result<(State, Option<usize>), SessionError> {
-        let mut len = None;
-        session.need_with::<AnonymousToken, _, ()>(&EmptyProvider, |token| {
-            let buf = token.as_bytes();
-            writer.write_all(buf)?;
-            len = Some(buf.len());
+        let len = session.maybe_need_with::<AnonymousToken, _, _>(&EmptyProvider, |token| {
+            writer.write_all(token.as_bytes())?;
+            Ok(token.len())
+        })?.unwrap_or(0);
+        Ok((State::Finished, Some(len)))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::io::Cursor;
+    use crate::callback::{Context, Request, SessionCallback, SessionData};
+    use crate::error::SessionError;
+    use crate::mechanisms::anonymous::AnonymousToken;
+    use crate::test;
+
+    struct C<'a> {
+        token: Option<&'a str>,
+    }
+    impl SessionCallback for C<'_> {
+        fn callback(
+            &self,
+            _session_data: &SessionData,
+            context: &Context,
+            request: &mut Request,
+        ) -> Result<(), SessionError> {
+            if let Some(token) = self.token {
+                request.satisfy::<AnonymousToken>(token)?;
+            }
+
             Ok(())
-        })?;
-        Ok((State::Finished, len))
+        }
+    }
+
+    fn test_token(token: Option<&'static str>, output: &[u8]) {
+        let config = test::server_config(C { token });
+        let mut session = test::client_session(config, &super::super::mechinfo::ANONYMOUS);
+        let mut out = Cursor::new(Vec::new());
+
+        let (state, written) = session.step(None, &mut out).unwrap();
+
+        let data = out.into_inner();
+        assert_eq!(written, Some(data.len()));
+        assert_eq!(output, &data[..]);
+        assert!(state.is_finished());
+    }
+
+    #[test]
+    fn test_with_token() {
+        let tokens = [
+            "thisisatoken",
+            "", // empty string should not error.
+        ];
+        for token in tokens {
+            test_token(Some(token), token.as_bytes());
+        }
+    }
+
+    #[test]
+    fn test_no_token() {
+        test_token(None, b"");
     }
 }