+++ CHANGELOG.md
@@ -18,6 +18,7 @@ release. They will however *never* happen in a patch release.
 
 ## Fixed
 - `ANONYMOUS` client now correctly allows no token to be provided
+- `ANONYMOUS` server will now correctly allow no token to be provided by a client
 
 ## [v2.0.0-preview9] — 2022-08-08