+++ CHANGELOG.md
@@ -22,6 +22,7 @@ release. They will however *never* happen in a patch release.
 - `EXTERNAL` client now correctly requests an optional `AuthzId` instead of a required `AuthId`
 - `EXTERNAL` server now correctly provides `Authzid` instead of `Authid`
 - `PLAIN` client correctly rejects empty authid or password. Empty authzid is still treated as no authzid provided.
+- `PLAIN` server now allows non-UTF8 passwords, and does not apply saslprep to the authzid anymore.
 
 ## [v2.0.0-preview9] — 2022-08-08