+++ CHANGELOG.md
@@ -16,6 +16,18 @@ release. They will however *never* happen in a patch release.
 <!-- next-header -->
 ## [Unreleased] — ReleaseDate
 
+## [v2.0.0-preview10] — 2022-08-24
+
+## Fixed
+- `ANONYMOUS` client now correctly allows no token to be provided
+- `ANONYMOUS` server will now correctly allow no token to be provided by a client
+- `EXTERNAL` client now correctly requests an optional `AuthzId` instead of a required `AuthId`
+- `EXTERNAL` server now correctly provides `Authzid` instead of `Authid`
+- `PLAIN` client correctly rejects empty authid or password. Empty authzid is still treated as no authzid provided.
+- `PLAIN` server now allows non-UTF8 passwords, and does not apply saslprep to the authzid anymore. If authid or 
+  password are empty after saslprep an error is returned.
+- `SCRAM` server now calls validate correctly
+
 ## [v2.0.0-preview9] — 2022-08-08
 
 ### Changed
@@ -88,7 +100,8 @@ making this release possible.**
 * Initial stabilized release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/dequbed/rsasl/compare/v2.0.0-preview9...development
+[Unreleased]: https://github.com/dequbed/rsasl/compare/v2.0.0-preview10...development
+[v2.0.0-preview10]: https://github.com/dequbed/rsasl/compare/v2.0.0-preview9...v2.0.0-preview10
 [v2.0.0-preview9]: https://github.com/dequbed/rsasl/compare/v2.0.0-preview8...v2.0.0-preview9
 [v2.0.0-preview8]: https://github.com/dequbed/rsasl/compare/v2.0.0-preview7...v2.0.0-preview8
 [v2.0.0-preview7]: https://github.com/dequbed/rsasl/compare/v2.0.0-preview1...v2.0.0-preview7